+++ iotedgehubdev/__init__.py
@@ -1,12 +1,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 
-
-import pkg_resources
-
-pkg_resources.declare_namespace(__name__)
-
 __author__ = 'Microsoft Corporation'
-__version__ = '0.14.18'
+__version__ = '0.14.19'
 __AIkey__ = '95b20d64-f54f-4de3-8ad5-165a75a6c6fe'
 __production__ = 'iotedgehubdev'