+++ tests/test_compose.py
@@ -4,6 +4,7 @@
 
 import json
 import os
+import posixpath
 import shutil
 
 import pytest
@@ -302,7 +303,7 @@ def test_bind(source, target, mode):
     volume = {
         'source': source,
         'target': target,
-        'type': 'bind' if source is not None and os.path.isabs(source) else 'volume'
+        'type': 'bind' if source is not None and (os.path.isabs(source) or posixpath.isabs(source)) else 'volume'
     }
 
     if mode == 'ro':
@@ -375,18 +376,20 @@ def test_join_create_options():
         "createOptions01": ": {'PortBindings': {'43/udp': [{'HostPort': '4",
         "createOptions02": "3'}], '42/tcp': [{'HostPort': '42'}]}}}"
     }''')
-    assert ComposeProject._join_create_options(valid_settings) == """{'Env': ['k1=v1', 'k2=v2', 'k3=v3'], 'HostConfig': {'PortBindings'\
+    assert ComposeProject._join_create_options(valid_settings) == (
+        """{'Env': ['k1=v1', 'k2=v2', 'k3=v3'], 'HostConfig': {'PortBindings'\
 : {'43/udp': [{'HostPort': '43'}], '42/tcp': \
-[{'HostPort': '42'}]}}}"""
+[{'HostPort': '42'}]}}}""")
 
     valid_settings_unsorted = json.loads('''{
         "createOptions01": ": {'PortBindings': {'43/udp': [{'HostPort': '4",
         "createOptions02": "3'}], '42/tcp': [{'HostPort': '42'}]}}}",
         "createOptions": "{'Env': ['k1=v1', 'k2=v2', 'k3=v3'], 'HostConfig'"
         }''')
-    assert ComposeProject._join_create_options(valid_settings_unsorted) == """{'Env': ['k1=v1', 'k2=v2', 'k3=v3'], 'HostConfig': {'PortBindings'\
+    assert ComposeProject._join_create_options(valid_settings_unsorted) == (
+        """{'Env': ['k1=v1', 'k2=v2', 'k3=v3'], 'HostConfig': {'PortBindings'\
 : {'43/udp': [{'HostPort': '43'}], '42/tcp': \
-[{'HostPort': '42'}]}}}"""
+[{'HostPort': '42'}]}}}""")
 
     valid_settings_full = json.loads('''{
         "createOptions": "{'Env': ['k1=v1', 'k2=v2', 'k3=v3'], 'HostConfig'",
@@ -398,9 +401,10 @@ def test_join_create_options():
         "createOptions06": "}",
         "createOptions07": "}"
     }''')
-    assert ComposeProject._join_create_options(valid_settings_full) == """{'Env': ['k1=v1', 'k2=v2', 'k3=v3'], 'HostConfig': {'PortBindings'\
+    assert ComposeProject._join_create_options(valid_settings_full) == (
+        """{'Env': ['k1=v1', 'k2=v2', 'k3=v3'], 'HostConfig': {'PortBindings'\
 : {'43/udp': [{'HostPort': '43'}], '42/tcp': \
-[{'HostPort': '42'}]}}}"""
+[{'HostPort': '42'}]}}}""")
 
     invalid_settings_1 = json.loads('''{
         "createOptions": "{'Env': ['k1=v1', 'k2=v2', 'k3=v3'], 'HostConfig'",